Applied Requested Changes
